Minor styling changes
