+++ src/components/Button/styles.ts
@@ -11,6 +11,7 @@ const ButtonWrapper = styled.button`
 
   &.mirror > img {
     transform: rotate(180deg);
+    
   }
 `;
