+++ src/components/Button/types.ts
@@ -1,0 +1,8 @@
+export interface IButtonAttributes {
+    className?: string;
+    description?: string;
+    src?: string;
+    alt?: string;
+    action?: string;
+    onClick?: () => void;
+  }
