--- src/components/ConfirmBox/ConfirmBox.style.ts
@@ -1,20 +1,0 @@
-import styled from "styled-components";
-
-const ConfirmBoxWrapper = styled.div`
-  padding: 10px;
-  .confirm-box-buttons {
-    margin: 0 0 10px 350px;
-    gap: 15px;
-
-    button {
-      width: 60px;
-      justify-content: center;
-    }
-  }
-
-  & span.negative {
-    color: red;
-  }
-`;
-
-export default ConfirmBoxWrapper;
