--- src/components/Footer/FooterWrapper.tsx
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-const FooterWrapper = styled.footer`
-  align-items: center;
-  justify-content: center;
-`;
-export default FooterWrapper;
