--- src/components/Header/Header.style.ts
@@ -1,16 +1,0 @@
-import styled from "styled-components";
-
-const HeaderWrapper = styled.header`
-  height: 80px;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 20px;
-  margin-top: 20px;
-  align-items: center;
-
-  .logo {
-    width: 150px;
-  }
-`;
-
-export default HeaderWrapper;
