--- src/components/Pagination/Pagination.style.ts
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-const PaginationWrapper = styled.section`
-  align-items: center;
-  justify-content: center;
-`;
-
-export default PaginationWrapper;
