--- src/components/SearchBar/SearchBar.style.ts
@@ -1,13 +1,0 @@
-import styled from "styled-components";
-const SearchBarWrapper = styled.div`
-  align-items: center;
-  background-color: var(--search-bar);
-  border-radius: 5px;
-  padding: 5px;
-  input {
-    border: none;
-    height: 35px;
-    border-radius: 5px;
-  }
-`;
-export default SearchBarWrapper;
