+++ src/components/Table/Table.tsx
@@ -6,7 +6,7 @@ import sortIcon from "../../assets/images/ascending_order_icon.svg";
 import TableWrapper from "./styles";
 import { Link } from "react-router-dom";
 import { employees, tableHeaders } from "../../core/config/constants";
-// import { IRowProps } from "../../core/interface/interface";
+// import { IRowProps } from "./types";
 
 const TableHeader = ({
   tableHeader,
