+++ src/components/Table/types.ts
@@ -1,0 +1,6 @@
+// export interface IRowProps {
+//   id: string;
+//   employeeName: string;
+//   department: string;
+//   role: string;
+// }
