--- src/components/TableOptions/TableOptions.style.ts
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-const TableOptionsWapper = styled.section`
-  justify-content: space-between;
-  align-items: center;
-`;
-
-export default TableOptionsWapper;
