--- src/core/interface/interface.ts
@@ -1,17 +1,0 @@
-export interface IButtonAttributes {
-  className?: string;
-  description?: string;
-  src?: string;
-  alt?: string;
-  action?: string;
-  onClick?: () => void;
-}
-
-// export interface IRowProps {
-//   id: string;
-//   employeeName: string;
-//   department: string;
-//   role: string;
-// }
-
-// export const tableHeaders = ["Employee Number", "Name", "Department", "Role"];
