+++ src/pages/ViewEmployee/styles.ts
@@ -1,11 +1,7 @@
 import styled from "styled-components";
 
 const DivWrapper = styled.div`
-  padding: 15px;
-  background-color: var(--table);
-  border-radius: 5px;
-
-  h2{
+  h2 {
     margin-top: 0;
   }
 `;
